backend/requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
PyYAML==3.11
